These are generated

+++ demo/exchanger/typechain/ForexFactory.ts
@@ -33,7 +33,7 @@ const _abi = [
       },
       {
         name: "_rate",
-        type: "int128"
+        type: "uint256"
       }
     ],
     name: "setRate",
@@ -87,7 +87,7 @@ const _abi = [
     outputs: [
       {
         name: "",
-        type: "int128"
+        type: "uint256"
       }
     ],
     payable: false,
